Modified based on comments
