+++ python/emach/model_obj/Dimensions/dim_base.py
@@ -1,9 +1,26 @@
 
-class DimBase(float):
+from abc import abstractmethod, ABC
+
+
+class DimBase(float, ABC):
     def __new__(cls, value):
         return float.__new__(cls, value)
     def __init__(self,data):
         self.data = float(data)
+
+    @property
+    @abstractmethod
+    def conversion_factor(self):
+        pass
+
+
+    def _to_dimensionless(self):
+        return float(self.data * self.conversion_factor)
+
+
+    def _from_dimensionless(self):
+        x = self.result / self.conversion_factor
+        return self.__class__(x)
