+++ python/emach/model_obj/Dimensions/dim_millimeter.py
@@ -4,15 +4,7 @@ __all__ = ['DimMillimeter']
 
 class DimMillimeter(DimLinear):
     def __new__(cls, value):
-        if type(value).__bases__[0] is DimLinear:
-            value = value.to_mm()
         return DimLinear.__new__(cls, value)
-
-    def to_mm(self):
-        return self
-
-    def to_inch(self):
-        return self / 25.4
 
     @property
     def conversion_factor(self):
